fix code
change type check logic in the function 'checkNumber'

+++ lista_1/1.py
@@ -1,7 +1,7 @@
 # Escreva um programa que receba um número e verifique se ele é positivo, negativo ou zero.
 
 def checkNumber(n):
-    if(type(n) != int and type(n) != int):
+    if(type(n) != int and type(n) != float):
         return TypeError
     if(n < 0):
         return "n < 0"
@@ -17,6 +17,3 @@ print(checkNumber("10"))
 # Mudanças recomendadas:
 # - Função isinstance(value, type) para checagem de tipo;
 # - Uso de else ao fim do algoritmo;
-
-
-
